packagesCommit
1.Resourses are moved to resourses package
2. Models are moved to models package
3. Code is functional at this point
