+++ models/user.py
@@ -1,7 +1,6 @@
 import sqlite3
-from flask_restful import Resource, reqparse
 
-class User:
+class UserModel:
     def __init__(self, _id, username, password):
         self.id = _id
         self.password = password
@@ -40,35 +39,3 @@ class User:
 
         connection.close()
         return user
-
-class UserRegister(Resource):
-    parser = reqparse.RequestParser()
-
-    parser.add_argument(
-        'username',
-        type=str,
-        required=True,
-        help="This field cannot be empty"
-    )
-
-    parser.add_argument(
-        'password',
-        type=str,
-        required=True,
-        help="This field cannot be empty"
-    )
-
-    def post(self):
-        data = UserRegister.parser.parse_args()
-        if User.find_by_username(data['username']):
-            return {"message": "User " + data['username'] + " already exists"}, 400
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-
-        insert_query = "INSERT INTO users VALUES (NULL, ?, ?)"
-        cursor.execute(insert_query, (data['username'], data['password']))
-
-        connection.commit()
-        connection.close()
-
-        return {'message': 'User created successfully'}, 201
