+++ resourses/user.py
@@ -1,0 +1,36 @@
+import sqlite3
+from flask_restful import Resource, reqparse
+from models.user import UserModel
+
+
+class UserRegister(Resource):
+    parser = reqparse.RequestParser()
+
+    parser.add_argument(
+        'username',
+        type=str,
+        required=True,
+        help="This field cannot be empty"
+    )
+
+    parser.add_argument(
+        'password',
+        type=str,
+        required=True,
+        help="This field cannot be empty"
+    )
+
+    def post(self):
+        data = UserRegister.parser.parse_args()
+        if UserModel.find_by_username(data['username']):
+            return {"message": "User " + data['username'] + " already exists"}, 400
+        connection = sqlite3.connect('data.db')
+        cursor = connection.cursor()
+
+        insert_query = "INSERT INTO users VALUES (NULL, ?, ?)"
+        cursor.execute(insert_query, (data['username'], data['password']))
+
+        connection.commit()
+        connection.close()
+
+        return {'message': 'User created successfully'}, 201
